This is the template for the ppm-8 project. It was designed for the Milestone 1 documentation. It is basically a template for project requirements. It has images from an example project deployed with SparxSystems Enterprise Architect.
This document will mature over time until the Deadline of Milestone 1. Once it is finished, the document will be frozen and a final copy will be published at the Documents and Files section of the website. This version in the repository will be removed to avoid duplicates and only the final draft will be available on the downloads section.