Sections of the Miltstone 1 documentation.
Each project member will be working freely with the template in the repository added on last revision (INFINITYNET_M1_REQSPEC.doc). However, the additions to YOUR LOCAL COPY must be checked out into Subversion inside one of the *.txt file section committed in this revision. There will be daily iterations over the final document (.doc) in the end of the day to collect all the changes. This is necessary to avoid multiple changes to the MS Word document style. Images references can be done on the documentation as well, since they will be created along the development of the main document.