[Issue 3] Once again changing the classes directory structure to support the including of other components into the application.
- classes/infinitymetrics: the package for importing files will start with just infinitymetrics. In this way, a requires_once statement would be just as follows:

require_once("infinitymetrics/model/user/User.class.php");

Basically, this requirements comes with the complete name of the framework we are using, that is, in our case, infinitymetrics.

Just moved all the model/view/util etc to infinitymetrics. In this way, we can allow other components to be placed into the same hierarchy.

/classes/infinitymetrics
/classes/orm -> would be the ORM system that is used in parallel with infinitymetrics.
/classes/other_framework -> something else we can define.