[Issue 105]
Committing updated unit tests for EventChannel and Report testing new methods getEventsByUser and filterByUser respectively. Test fails as of this commit.

Note: filterByCategory follows the same procedure as the two tests above and passes test. We believe this may be due to the fact that the data member in EventCategory is a string and the data members of EventChannel and Event are objects, so the equality test using the == operator may fail when comparing objects whose data members are objects.