[Issue 117]
General Notes: Since the login is not complete as of the time of this commit, we used mock user objects (which are actually saved to the database) instead of using PHP session to look for the user_id of the person currently logged in. Once that is implemented, we can modify these files to make use of the PHP session.

Also, it is worth noting that the include paths for the header & footer DON'T work with files outside the 'template' root because it looks for the css in the current directory (e.g. when inside the 'workspace' directory, it looks for the css & other resources there instead of at the root of 'template' where they are located). Consquently, a script is necessary to generate these include paths correctly that will take into account the current location of the PHP webpage being displayed and the relative path back to the 'template' root. To overcome this for testing purposes, we simply placed all our PHP webpages at the root of 'template'. We are committing to the correct 'workspace' directory to maintain the correct directory structure, but if we don't get the sessions set up for this weekend's walk-through (which is fine), these files need to be moved to the root of 'template' to be executed properly.

workspaceCollection.php: We envisioned this page as being the homepage for the Workspace nav-button. It displays the collecion of Workspaces (both owned and shared with the user). Every item has a hyperlink that points to the viewWorkspace page for that workspace.

viewWorkspace.php: This page takes the workspace_id as a parameter encoded in the URL in order to display the details for the workspace.

createWorkspace.php: This page allows the user to create a new Workspace. PLEASE NOTE that the java.net parent project functionality to is NOT implemented. Will need to discuss with Marcello in order to integrate this portion with the Personal Agent. For now, the field is there and the functionality can be coded when ready.

shareWorkspace.php: The page shares a Workspace with another java.net user. The user must arrive at this page by pressing the 'Share Workspace' button on the individual workspace's viewWorkspace page. We are still missing a piece of implementation for this piece which is exception handling for workspaces that are already shared with a particular jn_username. We will be updating the controller tomorrow to account for this scenario.

Please contact me with any questions.