[Issue 126]
Updated getProjects method. The doSelect statement in the previous version returned an array an assigned to the $project variable. Therefore using the getParentProjectJnName on this variable failed as it was not a project object but an array of project objects (even if the query returned only one project in the array). The retrieveByPk guarantees that only one object is returned and stored in $project.