Fixed Add Entry bug that put new entries on the last created event.
Restructured old Delete Entry view to become a Delete Entry confirmation.

Replaced all blue-text hyperlink buttons with the same kind of buttons in the rest of the software.

Created a view specifically for seeing the contents and date of an individual custom event entry.

Simplified button structure in original View Custom Events in workspace view.